+++ lib/src/logic/platform_helper.dart
@@ -1,0 +1,6 @@
+import "package:flutter/cupertino.dart";
+import "package:flutter/material.dart";
+
+bool isMaterialApp(BuildContext context) => context.findAncestorWidgetOfExactType<MaterialApp>() != null;
+
+bool isCupertinoApp(BuildContext context) => context.findAncestorWidgetOfExactType<CupertinoApp>() != null;